Make vending optional. Extract interface out of impl
