+++ core/src/main/resources/reference.conf
@@ -574,7 +574,7 @@ services {
   GithubAuthVending {
     class = "cromwell.services.auth.impl.GithubAuthVendingActor"
     config {
-      access-token = "dummy-token"
+      enabled = false
     }
   }
 }
