+++ services/src/main/scala/cromwell/services/auth/GithubAuthVendingSupport.scala
@@ -1,0 +1,28 @@
+package cromwell.services.auth
+
+import akka.actor.ActorRef
+import akka.pattern.AskSupport
+import akka.util.Timeout
+import cromwell.languages.util.ImportResolver.ImportAuthProvider
+import cromwell.services.auth.GithubAuthVending.{GithubAuthRequest, GithubAuthVendingFailure, GithubAuthVendingResponse, GithubAuthTokenResponse}
+
+import scala.concurrent.{ExecutionContext, Future}
+
+trait GithubAuthVendingSupport extends AskSupport {
+
+  def serviceRegistryActor: ActorRef
+
+  implicit val timeout: Timeout
+  implicit val ec: ExecutionContext
+
+  def importAuthProvider(token: String): ImportAuthProvider = new ImportAuthProvider {
+    override def validHosts: List[String] = List("github.com", "githubusercontent.com", "raw.githubusercontent.com")
+
+    override def authHeader(): Future[Map[String, String]] = {
+      serviceRegistryActor.ask(GithubAuthRequest(token)).mapTo[GithubAuthVendingResponse].flatMap {
+        case GithubAuthTokenResponse(token) => Future.successful(Map("Authorization" -> s"Bearer ${token}"))
+        case GithubAuthVendingFailure(error) => Future.failed(error)
+      }
+    }
+  }
+}
